Update service client

+++ src/main/java/com/pragma/service/impl/ClientServiceImpl.java
@@ -1,6 +1,7 @@
 package com.pragma.service.impl;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import org.slf4j.Logger;
@@ -9,40 +10,44 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.pragma.entity.Client;
+import com.pragma.entity.ImageMongoDB;
 import com.pragma.entity.validate.ClientValidate;
 import com.pragma.repository.ClientRepository;
 import com.pragma.service.ClientService;
+import com.pragma.service.ImageMongoDBService;
 import com.pragma.util.Pragma;
 import com.pragma.util.exception.PragmaException;
 
-
 @Service
 public class ClientServiceImpl implements ClientService {
-	
+
 	private static final Logger LOGGER = LoggerFactory.getLogger(ClientServiceImpl.class);
 
 	@Autowired
 	ClientRepository clientEntityRepository;
-	
+
+	@Autowired
+	ImageMongoDBService imageMongoDBService;
+
 	public ClientServiceImpl(ClientRepository clientEntityRepository) {
 		this.clientEntityRepository = clientEntityRepository;
 	}
 
 	@Override
 	public Client findById(Long id) {
-		if(!Pragma.isLong(id))
+		if (!Pragma.isLong(id))
 			throw new PragmaException("El id del cliente no es valido.");
 		Optional<Client> optional = clientEntityRepository.findById(id);
-		if(!optional.isPresent())
-			throw new PragmaException("No se ha encontrado ningun cliente con el id "+id+".");
+		if (!optional.isPresent())
+			throw new PragmaException("No se ha encontrado ningun cliente con el id " + id + ".");
 		return optional.get();
 	}
 
 	@Override
 	public Client findByTypeAndDocument(String type, Long document) {
-		if(!Pragma.isString(type))
+		if (!Pragma.isString(type))
 			throw new PragmaException("El tipo de documento del cliente no es valido.");
-		if(!Pragma.isLong(document))
+		if (!Pragma.isLong(document))
 			throw new PragmaException("El documento del cliente no es valido.");
 		Client client = clientEntityRepository.findByTypeAndDocument(type, document);
 		if (client == null)
@@ -53,11 +58,11 @@ public class ClientServiceImpl implements ClientService {
 
 	@Override
 	public List<Client> findByHigherOrEqualsAge(int age) {
-		if(!Pragma.isInteger(age))
+		if (!Pragma.isInteger(age))
 			throw new PragmaException("La edad del cliente no es valido.");
 		return clientEntityRepository.findByHigherOrEqualsAge(age);
 	}
-	
+
 	@Override
 	public List<Client> findAll() {
 		return clientEntityRepository.findAll();
@@ -65,7 +70,7 @@ public class ClientServiceImpl implements ClientService {
 
 	@Override
 	public List<Client> findByType(String type) {
-		if(!Pragma.isString(type))
+		if (!Pragma.isString(type))
 			throw new PragmaException("El tipo de documento del cliente no es valido.");
 		return clientEntityRepository.findByType(type);
 	}
@@ -74,10 +79,10 @@ public class ClientServiceImpl implements ClientService {
 	public Client save(Client cliente) {
 		ClientValidate.message(cliente);
 		cliente.setId(0L);
-		if(!testTypeDocument(cliente.getType(), cliente.getDocument()))
+		if (!testTypeDocument(cliente.getType(), cliente.getDocument()))
 			throw new PragmaException("Ya existe un cliente con ese documento y tipo de documento.");
-		cliente= clientEntityRepository.save(cliente);
-		if(cliente == null)
+		cliente = clientEntityRepository.save(cliente);
+		if (cliente == null)
 			throw new PragmaException("No se ha registrado el cliente.");
 		return cliente;
 	}
@@ -86,11 +91,14 @@ public class ClientServiceImpl implements ClientService {
 	public Client update(Client cliente) {
 		ClientValidate.message(cliente);
 		Client aux = findById(cliente.getId());
-		if(!aux.getType().equalsIgnoreCase(cliente.getType()) || aux.getDocument() != cliente.getDocument())
-			if(!testTypeDocument(cliente.getType(), cliente.getDocument()))
+		System.out.println(cliente);
+		System.out.println(aux);
+		if (!aux.getType().equalsIgnoreCase(cliente.getType())
+				|| !Objects.equals(aux.getDocument(), cliente.getDocument()))
+			if (!testTypeDocument(cliente.getType(), cliente.getDocument()))
 				throw new PragmaException("Ya existe un cliente con ese documento y tipo de documento.");
 		cliente = clientEntityRepository.save(cliente);
-		if(cliente == null)
+		if (cliente == null)
 			throw new PragmaException("No se ha actualizado el cliente.");
 		return cliente;
 	}
@@ -98,27 +106,50 @@ public class ClientServiceImpl implements ClientService {
 	@Override
 	public boolean delete(String type, Long document) {
 		Client client = findByTypeAndDocument(type, document);
+		List<ImageMongoDB> list = imageMongoDBService.findByClient(client.getId());
+		if (Pragma.isList(list))
+			throw new PragmaException("No se ha eliminado el cliente con tipo de documento " + type + " y documento "
+					+ document + ", tiene asociado " + list.size() + " fotos.");
 		clientEntityRepository.deleteById(client.getId());
 		try {
 			client = findByTypeAndDocument(type, document);
-		}catch (PragmaException e) {
+		} catch (PragmaException e) {
 			LOGGER.error("delete(String type, Long document)", e);
 			client = null;
-		}finally {
-			if(client == null)
+		} finally {
+			if (client == null)
 				return true;
 		}
 		throw new PragmaException("No se ha eliminado el cliente.");
 	}
-	
+
+	@Override
+	public boolean delete(Long id) {
+		Client client = findById(id);
+		List<ImageMongoDB> list = imageMongoDBService.findByClient(client.getId());
+		if (Pragma.isList(list))
+			throw new PragmaException(
+					"No se ha eliminado el cliente con el id " + id + ", tiene asociado " + list.size() + " fotos.");
+		clientEntityRepository.deleteById(client.getId());
+		try {
+			client = findById(id);
+		} catch (PragmaException e) {
+			LOGGER.error("delete(String type, Long document)", e);
+			client = null;
+		} finally {
+			if (client == null)
+				return true;
+		}
+		throw new PragmaException("No se ha eliminado el cliente.");
+	}
+
 	private boolean testTypeDocument(String type, Long document) {
 		Client client = null;
 		try {
 			client = findByTypeAndDocument(type, document);
-		}catch (PragmaException e) {
+		} catch (PragmaException e) {
 			LOGGER.error("testTypeDocument(String type, Long document)", e);
 		}
 		return client == null;
 	}
 }
-
